uae-finance-dashboard/requirements.txt
streamlit
fastapi
uvicorn
pandas
google-generativeai
requests
python-dotenv
pydantic
